Unfold types into multiple files as originally designed
